Utilize comment's _fast_permalink attribute
Due to namespace collision the necessitated PRAW 3.6.2, the property
`permalink` was removed from comments. This works fine from comments obtained
as part of submission because they now have a first-class `permalink`
attribute. However, comments that are returned out of the inbox, do not come
with such an attribute. Unfortunately the same fix that enabled `permalink` to
work on the aforementioned comments, also results in PRAW no longer having a
`permalink` attribute for comments from the inbox.

Ref: https://github.com/praw-dev/praw/pull/860